feat: setup and teardown imrovement
