--- src/helpers/contract_types.ts
@@ -1,4 +1,0 @@
-export type AcknowledgementResult =
-  | { success: string[] }
-  | { error: string[] }
-  | { timeout: string };
