+++ src/testcases/interchain_tx_query.test.ts
@@ -1,9 +1,4 @@
-import {
-  COSMOS_DENOM,
-  CosmosWrapper,
-  NEUTRON_DENOM,
-  NeutronContract,
-} from '../helpers/cosmos';
+import { COSMOS_DENOM, CosmosWrapper, NEUTRON_DENOM } from '../helpers/cosmos';
 import { proto } from '@cosmos-client/core';
 import { TestStateLocalCosmosTestNet } from './common_localcosmosnet';
 import Long from 'long';
@@ -14,6 +9,7 @@ import {
   registerTransfersQuery,
   waitForTransfersAmount,
 } from '../helpers/icq';
+import { NeutronContract } from '../helpers/types';
 
 describe('Neutron / Interchain TX Query', () => {
   let testState: TestStateLocalCosmosTestNet;
